Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-works-on-frame-src.https.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'report-to'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'report-to'.

CONSOLE MESSAGE: Refused to load https://localhost:9443/content-security-policy/support/fail.html because it does not appear in the frame-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

TIMEOUT Event is fired Test timed out
FAIL Violation report status OK. assert_equals: No such report. expected "" but got "false"
